wip: create scaffolding for another rewrite
I realized that the last fix introduced a new problem where you could ONLY cancel the first level of a building, and canceling any OTHER levels would register as completion.

As I dug in I realized the issue was that none of the variables I was checking would update until 1 day after construction ended. So now all the logic takes places in an_action and I'll need to expand out the new structure to accomadate all the other foreign buildings.

As well as remove testing variables and update documentation.